****************CHANGES FROM MASTER************** - fixed obstacle removal and graphics glitches

+++ core/src/com/mygdx/game/objects/Overlay.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.graphics.Texture;
  * Created by Samuel on 21/3/2016.
  */
 public class Overlay extends Movable{
-    public Overlay(int y){
+    public Overlay(float y){
         super(0, y, 480, 800);
         this.setImage(new Texture(Gdx.files.internal("effects1.png")));
     }
